refactor: generateCTSchedulesをmember.tsに切り出し
